+++ src/main/kotlin/de/andrena/tools/altn8th/domain/relatedFiles/find/strategies/prefix/FindRelatedFilesByPrefixStrategy.kt
@@ -30,5 +30,5 @@ internal class FindRelatedFilesByPrefixStrategy : FindRelatedFilesStrategy {
             || areRelatedByGivenPattern(relatedFile, basename, relatedFileHop.pattern)
 
     private fun areRelatedByGivenPattern(first: String, second: String, prefixPattern: String): Boolean =
-        first.matches(Regex("^(${prefixPattern})${second}$"))
+        first.matches(Regex("^(${prefixPattern})${second}$", RegexOption.IGNORE_CASE))
 }