+++ CHANGELOG.md
@@ -7,12 +7,16 @@ Navigate to related files using a shortcut, which is (Alt)(8) by default.
 ### Added
 
 ### Changed
+- Prefix match case-insensitively
+- Postfix match case-insensitively
+- Free relations match case-insensitively
+- Files with same filename but different extension match case-insensitively
 
 ### Fixed
 
 ### Removed
 
-## [1.0.0] - 2024-06-08
+## 1.0.0 - 2024-06-08
 
 ### Fixed